Version 2.0
Major changes were made in order to make this code more robust.  The
most notable change was converting the JavaScript function to utilize
an object, and converting the object to JSON for transferring it to the
server for reading.  This change alone has resulted in code that is far
cleaner, leaner, and stronger than it was before for both the
JavaScript and back-end PHP file.  Another change was to dynamically
generate the table to be a 10 columns and 10,000 rows to demonstrate
the cabability of the code.